[Refactor]-Cart 관련 기능 코드 개선, 전역 상태 제거
