+++ src/lib/cart/hooks/useFetchCart.ts
@@ -1,15 +1,29 @@
 import { useAuthStore } from "@/store/auth/useAuthStore";
 import { useQuery } from "@tanstack/react-query";
 import { fetchCartItems } from "../api";
+import { CartItem } from "../types";
 
 export const useFetchCart = () => {
   const { user } = useAuthStore();
 
-  return useQuery({
+  const {
+    data: cartItems = [],
+    isLoading,
+    error,
+  } = useQuery<CartItem[]>({
     queryKey: ["cartItems", user?.uid],
-    queryFn: () => {
-      return fetchCartItems(user!);
-    },
+    queryFn: () => fetchCartItems(user!),
     enabled: !!user,
   });
+
+  const totalQuantity = cartItems.reduce(
+    (total, item) => total + item.quantity,
+    0
+  );
+  const totalPrice = cartItems.reduce(
+    (total, item) => total + item.productPrice * item.quantity,
+    0
+  );
+
+  return { cartItems, totalQuantity, totalPrice, isLoading, error };
 };
