+++ src/lib/cart/types.ts
@@ -1,0 +1,5 @@
+import { Product } from "../product/types";
+
+export interface CartItem extends Product {
+  quantity: number;
+}
